+++ src/core/models/logo/logo.ts
@@ -1,0 +1,21 @@
+import Base from '../../../base/base';
+
+export interface ILogo {
+
+    base64?: string;
+
+}
+
+export class Logo extends Base.BaseModel implements ILogo {
+
+    constructor(params: ILogo = null) {
+
+        super(params);
+
+        this.base64 = params ? params.base64 : "";
+
+      }
+
+      base64: string;
+
+}