+++ src/core/models/models.ts
@@ -2,6 +2,7 @@ import { Entreprise, IEntreprise } from './entreprise/entreprise';
 import { Adresse, IAdresse } from './adresse/adresse';
 import { Identification, IIdentification } from './identification/identification';
 import { Personne, PersonnePhysique, PersonneMorale, IPersonne, IPersonneMorale, IPersonnePhysique } from './personne/personne';
+import { Logo, ILogo } from './logo/logo';
 
 export{ 
     Entreprise, IEntreprise,
@@ -10,5 +11,6 @@ export{
     Personne,
     PersonneMorale,
     PersonnePhysique,
-    IPersonne, IPersonneMorale, IPersonnePhysique
+    IPersonne, IPersonneMorale, IPersonnePhysique,
+    Logo, ILogo
 };